flatten hierarchy of demos folder
