+++ packages/demos/package.json
@@ -1,17 +1,35 @@
 {
   "name": "demos",
-  "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "version": "1.0.0",
+  "private": true,
   "scripts": {
+    "build": "rollup -c",
+    "dev": "rollup -c -w",
+    "start": "sirv public --no-clear",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
     "test": "echo \"Error: no test specified\" && exit 1",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.0",
+    "rollup": "^2.3.4",
+    "rollup-plugin-css-only": "^3.1.0",
+    "rollup-plugin-livereload": "^2.0.0",
+    "rollup-plugin-svelte": "^7.0.0",
+    "rollup-plugin-terser": "^7.0.0",
+    "rollup-plugin-postcss": "^4.0.1",
+    "postcss": "^8.3.9",
+    "postcss-import": "^14.0.2",
+    "svelte-check": "^2.0.0",
+    "svelte-preprocess": "^4.0.0",
+    "@rollup/plugin-typescript": "^8.0.0",
+    "typescript": "^4.0.0",
+    "tslib": "^2.0.0",
+    "@tsconfig/svelte": "^2.0.0",
     "@babel/core": "^7.20.12",
     "@storybook/addon-actions": "^6.5.16",
     "@storybook/addon-docs": "^6.5.16",
@@ -29,16 +47,17 @@
     "svelte-loader": "^3.1.5"
   },
   "dependencies": {
+    "sirv-cli": "^2.0.0",
     "@types/plotly.js-dist-min": "^2.3.0",
     "@visahoi/plotly": "3.0.0",
     "@visahoi/echarts": "3.0.0",
     "@visahoi/vega": "3.0.0",
     "plotly.js": "^2.2.1",
-    "plotly.js-dist-min": "~2.12.0", 
+    "plotly.js-dist-min": "~2.12.0",
     "echarts": "^5.1.2",
     "vega": "5.21.0",
     "vega-embed": "6.19.1",
-    "vega-lite": "5.1.1",   
+    "vega-lite": "5.1.1",
     "svelte-resize-observer": "^2.0.0"
   }
 }
