+++ packages/demos/src/Components/Echarts/Barchart/Barchart.svelte
