+++ packages/demos/src/Components/Echarts/Scatterplot/Scatterplot.svelte
