+++ packages/demos/src/Components/Plotly/Barchart/barchart.stories.svelte
