+++ packages/demos/src/Components/Plotly/Barchart/barchart.svelte
