+++ packages/demos/src/Components/Plotly/Scatterplot/Scatterchart.svelte
