+++ packages/demos/src/Components/Vega-lite/Barchart/Barchart.svelte
