+++ packages/demos/src/Components/Vega-lite/Scatterplot/Scatterplot.svelte
